+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Viewport toolbar: added transform orientation selector (World/Local) with dedicated icons
 - Transform controls: Ctrl-key snapping
 - Material properties: Enviroment map settings
+- Material properties: vector2 inputs
 
 ### Changed